added  DELETE /api/expenses/{expenseId} API and updated add_expense API
memberBalances list is added to Groups table in database.
in expense delete and add operation, this memberBalances is updated.
In this commit, DELETE /api/expenses/{expenseId} API  is implemented from scratch. add_expense API is updated by adding update mechanism of memberBalances. Also delete expense api is hooked to frontend. User who is owner of the group or payer of the expense can delete the expense, other cannot.